ENH: change bad naming in callbacks
